[Refactor] (code) Move all enum objects in '.model.metadata' to new module '.model.metadata_enum'.
