+++ smoothcrawler_cluster/model/__init__.py
@@ -1,1 +1,2 @@
-from .metadata import CrawlerStateRole, TaskResult, State, Task, Heartbeat
+from .metadata import State, Task, Heartbeat
+from .metadata_enum import CrawlerStateRole, TaskResult
