+++ smoothcrawler_cluster/model/metadata.py
@@ -1,69 +1,7 @@
-from datetime import datetime as dt
 from typing import List, Union
-from enum import Enum
+from datetime import datetime as dt
 
-
-
-class CrawlerStateRole(Enum):
-    """
-    This role is NOT the role of *SmoothCrawler-AppIntegration*. They're very different. The role in
-    *SmoothCrawler-AppIntegration* means source site (or producer) of application or processor site
-    (or consumer) of application. But the role in meta-data in *SmoothCrawler-Cluster* means it is
-    active runner to run task or backup of that active runner.
-
-    For **SmoothCrawler-Cluster** realm, it has 4 different roles:
-
-    * Runner
-    * Backup Runner
-    * Dead Runner
-    * Dead Backup Runner
-    """
-
-    Runner = "runner"
-    """ Literally, **Runner** role is the major element to run web spider tasks. """
-
-    Backup_Runner = "backup-runner"
-    """
-    **Backup Runner** role is the backup of **Runner**. It would activate (base on the setting, it may 
-    activate immediately) and run the web spider task if it still not finish. 
-    A **Backup Runner** would keep checking the heartbeat info of **Runner**, standby and ready to run 
-    in anytime for any one of **Runner** does not keep updating its own heartbeat info (and it would turn 
-    to **Dead Runner** at that time).
-    """
-
-    Dead_Runner = "dead-runner"
-    """ 
-    If **Runner** cannot work finely, like the entire VM be shutdown where the crawler runtime environment 
-    in. It would turn to be **Dead Runner** from **Runner**. In other words, it must to be **Dead Runner** 
-    if it cannot keep updating its own heartbeat info.
-    """
-
-    Dead_Backup_Runner = "dead-backup-runner"
-    """ **Dead Backup Runner** is same as **Dead Runner** but it's for **Backup Runner**. """
-
-
-
-class TaskResult(Enum):
-    """
-    The task result means it is the result of running web spider task. The web spider task could classify to
-    be 4 different states: processing, done, terminate and error.
-    """
-
-    Processing = "processing"
-    """ Task running in processing. """
-
-    Done = "done"
-    """ Finish the task and it works finely without any exceptions. """
-
-    Terminate = "terminate"
-    """ 
-    Web spider task running has been terminated so that it cannot finish all processes, but it doesn't occur 
-    any exceptions except KeyboardInterrupt.
-    """
-
-    Error = "error"
-    """ If it raise any exceptions in web spider task running, its result would be error. """
-
+from .metadata_enum import CrawlerStateRole, TaskResult
 
 
 class State:
